requirements.txt
Django~=5.1.1
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-bootstrap5~=24.3 
django-bootstrap-icons~=0.8.6
Pillow==9.4.0
django-environ~=0.11.2
djangorestframework~=3.14.0
